Move read_comms_env_vars to param/utility.py
